refactor: mark internal modules as such
