+++ src/safeds_runner/server/_json_encoder.py
