+++ tests/safeds_runner/server/test_json_encoder.py
@@ -6,7 +6,7 @@ from typing import Any
 import pytest
 from safeds.data.image.containers import Image
 from safeds.data.tabular.containers import Table
-from safeds_runner.server.json_encoder import SafeDsEncoder
+from safeds_runner.server._json_encoder import SafeDsEncoder
 
 
 @pytest.mark.parametrize(
